Added error checking for command-line arguments.
Common edge-case errors are now checked after each argument's ArgType
is determined. After each ArgType is determined, argt[] is looped
through, checking for errors.

Errors checked:

1. OUTPUT_FILE_FLAG provided, but not OUTPUT_FILE_NAME.
2. OUTPUT_FORMAT_FLAG provided, but not OUTPUT_FORMAT_NAME.
3. Multiple INPUT_FILE_NAME's provided.
    - This is currently the default for an unknown (or out of place)
      argument.
    - Perhaps it shouldn't be the default?
    - Then again, the default should imply that it would have the
      highest frequency of use, which is obviously the truth, since
      it is required.

A function was added, testArgTypeForMatch(), that tests an individual
argument against the required argument (based on its position in
argv[]). It returns nothing, but exits with warning dialogs if a
discrepancy is found.

A local variable was added to parseArguments(), inputFiles, which
keeps track of the number of arguments determined to be of type
INPUT_FILE_NAME. There is probably a better way to do this (perhaps
check this as we are assigning an ArgType of INPUT_FILE_NAME..), but
this is the current, simple, solution.