Add file processing, output for --help, --version
Added files.c and files.h. They contain functions that are called to
enable file processing (opening files for reading/writing). The plan
is to have all functions/structures related to file processing inside
this file (read from file/write to file). Also in the file, global
variables that contain the program name and version number, both used
by the various functions in this file; they are not exported.

Updated .gitignore to ignore `.DS_Store` as well as example input/
output files I'm using locally for testing.

Added a new enum in arguments.h, `formatTypes`, which contains the
valid types of OUPUT_FORMAT_NAME. Currently it contains HTML_FORMAT
and TXT_FORMAT. Neither are implemented, but it's a good starting
point.

Currently, main.c is parsing the arguments to determine each
arguments ArgType, then loop thought the ArgType array, `argt`. This
is how it is determining the function of the program. Perhaps a
better way will be determined in the future.

ANDALE ANDAL MOMMA E-I E-I OO OOOOOOHHHH