Added tests.c -> TIME TO GET TESTY
Added a file, tests.c, that implements tests for each of the major
functions of patdown. Currently, that means argument parsing and
file processing. Also, added support for printing with color codes,
albeit only in tests.c for now. Perhaps if the output is to stdin
the user could flag the program to tell it to syntax highlight it in
the shell??? GOD DREAM.

More importantly: Currently have to build tests with the following:
    clang arguments.c tests.c errors.c files.c

I tried the whole 'different folders' approach for a minute and
didn't like it, but I may go back to it.

Attempting to do a little test-driven development going forward. But
that is kind of a pipe dream. Time will tell.

Also added some more (explicit) documentation in files.c,
arguments.c, and errors.c. There are were getting to be a few too
many **magic** little bits of code that needed to be explained.

TODO: BUILD A GATTDAMN MAKEFILE.

THE PYRAMIDS SHALL RISE.