Added reading-from/writing-to files, format checking
The FormatType of each format string is now returned from
testFormatType() in arguments.c. This allows main() to easily
determine what should be assigned to outputFormat.

Added a input file error and a memory error message dialog.

And finally I added two new functions:

- char *readLine(FILE *fp)
- int writeLine(const char *line, FILE *fp)

readLine() allocates memory for one line, then reads it from *fp
using fgets(). It returns the string pulled from the file, or NULL
if it encountered an error / or EOF.  Probably could use more error
checking for EOF and whatnot here.

writeLine() uses fputs() to write a string to *fp. It returns a
boolean integer based on the success of the fputs() call.

In main(), I switched to now defaulting the intputFile to stdin and
the outputFile to stdout. I don't know if this is the best option,
but it makes testing considerably easier. I also laid out the basics
of the procedure for reading/processing/writing:

1. Consume all lines from intputFile
2. Process them using outputFormat
3. Write all lines to outputFile

All of this can happen during a while (1) loop that is broken out of
by the readLine() function.

Two new tests were added to tests.c that write a string to a file
then read that same string from that same file.