add tests for isATXHeader(): GET TOP ON THE PHONE
Added 14 tests for the isATXHeader() function.

Also laid the groundwork for how the tests for block-level elements
will be conducting in the future.  Next parser to be implemented
(horizontal rules probably) should be written WITH the tests, or
maybe I will write the tests first...

Also added one more edge-case check in the isATXHeader() function
that tests for padded spaces after the initial set of `#`s, for
example: '#            foo'. 12 spaces will be removed and when
the element string is allocated it will only contains 'foo'.

GET TOP ON THE PHONE.