Add HTML Blocks && code cleanup in buildList()
Removed the all the parsing logic from buildList(). This involved
created a new enumeration type `insert_t` that returns a value to
alert buildList() on how to handle the returning `temp_block_node_t`.

Added RUDIMENTARY SUPPORT for HTML Blocks and HTML Comments.

STILL NEED TO ADD TESTS FOR ALL THIS; IT HAS NOT BEEN EXTENSIVELY
TESTED. NOW THAT THE buildList() FUNCTION IS IN WORKING ORDER THE
NEXT COMMIT NEEDS TO BE A TEST SUITE UPDATE.