WAVES DONT DIE -> FIXES, FIXES, & REWRITES
For starters, the .gitignore file was updated to ignore .dYSM files
that are created by valgrind.  BTW, been using valgrind, FTW!!!

Since we're on valgrind, I went through are removed all the memory
leaks, aka: decided to free up some of that memory.  For now, all
all the memory leaks (that are my fault) are cleared up. Still
getting weird fucking errors that fgetc() is leaking, but I can't
see how, and from what I can tell this has something to do with the
library it uses to read files, not from me. This is one of the
reasons why every file shows its been MODIFIED.

Next, reworked by commenting style to be more unified across all the
files. Extended by wrap length to 75 columns (up from 70).  This
obviously does not include function definitions and headers, as most
of them are over 75 columns.

Next, switched up my naming conventions again, this is the most
unified they have been though, and if anyone else ever reads this,
it will make a fuckton more sense now. Here is where we are:
    - structs: typedef to a camel-case word where the first letter
               is always capitalized.
        - example: MarkdownBlock
    - enums: typedef to an all lowercase word appended with `_t`
        - example: mdblock_t
    - functions: start with verb, all lowercase words separated by
                 underscores `_`
        - example: alloc_parse_info()

Next, I factored out a lot of the repetitive create/alloc code. So
now all structures (and strings for that matter) are created
dynamically by calling a separate function with the initialization
values. That function returns the allocated structure, or exits if
malloc returns NULL.

The string functions above are placed in the new file utilities.c
which will house all these random ass functions that are helpful.

AND FINALLY: I completely rewrote the parsers.c file from scratch.
The only parser that is currently working is ATX_HEADERS, and it
is also completely rewritten. Honestly, I wasn't happy with the
previous attempt at it, and I am learning a lot of useful shit as
I go, and its motivating to rewrite these shit-ally implemented
functions.

Still need to rework the functions, comments, and naming in files.c,
but I reached a stopping point and figured I would cement this
update.

Also, obviously, all the tests are broken. And none of them have
been rewritten...