redid test suite -- added Link Reference parsing
================================================
TEST SUITE
================================================

The old test suite in patdown/tests.c has been deleted. Now all tests
are in patdown/tests/. The `make test` command has been updated to
create a new binary that determines which tests to run by the arguments
given at execution.

So far, there are three valid arguments:

1. `test parsers`
    - runs the parser suite in test_parsing.c
    - currently tests:
        - ATX_HEADING_x
        - HORIZONTAL_RULE
        - SETEXT_HEADING_x
        - INDENTED_CODE_BLOCK
2. `test all`
    - runs all tests
    - currently this is only the parsers
3. `test [all|parsers] -v`
    - `v` flag triggers *verbose* option
    - prints a short description of each test in addition to results

All the logic of the tests is in tests/tests.c, while the individual
tests for each segment of patdown will be located in test_<segment>.c
files as they are implemented.

================================================
FILES
================================================

The write_line() function has been updated to receive an additional
argument, newline, which determines if a newline should be printed
after printing all the strings passed as arguments.  This is a Boolean
value, and thus <stdbool.h> has been included in the source and header
file.

The new function definition is:

void write_line(FILE *fp, bool newline, size_t i, ...)

================================================
MAKEFILE
================================================

The new test suite has been implemented so that the `test` binary is
created in the main folder, even though the individual test files
are located in the tests/ subfolder.

May change this later to be more succinct.

================================================
MARKDOWN
================================================

A new data type, `LinkRef`, has been declared in the header to hold
information obtained from parsing a link reference definition. This
structure was designed to be implemented as a binary tree, so to
enable easy search to resolve the inline link references found.

A pointer to a `LinkRef` structure has also been added to the
`TempMarkdownBlock` structure, so that LinkRef's can be returned
from the parsing functions.

A new `insert_t` type has been defined as `INSERT_LINK`.  It alerts
the build_queue() function that this TempMarkdownBlock contains a
reference to a LinkRef, and instead of queueing this block, insert
it into the binary tree of link references.

Four functions have been defined to deal with the trees.  They are
**VERY TEMPORARY** in their design, and mostly implemented to confirm
that the tree is operating properly and the references are being
parsed correctly.  They are:

1. void insert_link_ref(LinkRef **tree, LinkRef *node);
2. LinkRef *alloc_link_ref(char *label, char *url, char *title);
3. void print_tree(LinkRef *currentNode);
4. void free_tree(LinkRef *currentNode);

Less importantly, I have begun the slow transition from checking
each new structure's existence against NULL to just checking its
existence. I think this is much more readable. And will be used
going forward.

To be clear this means no more of:

if (newNode != NULL) {...}

And, instead:

if (newNode) {...}

================================================
PARSERS
================================================

I finally gave up and implemented global variables in the parsers.c
source, mainly so that I am not passing 100 variables back and forth
between each of the parsing functions. So far, they are:

- lastBlock           --> mdblock_t of most recently parsed block.
- lastFencedBlockMark --> The mark used for most recent fenced code block.
- *linkLabelStr       --> A link label, used when parsing LinkRef's.
- *linkurlStr         --> A link URL, used when parsing LinkRef's.
- *linkTitleStr       --> A link title, used when parsing LinkRef's.

Going forward, I am going to be utilizing them more and more in an
attempt to only pass the line (read from input file) to each of the
parsing functions.

All parsing functions are no longer receiving the `last` argument, as
it is now a global variable.

Parsers have been added for FENCED_CODE_BLOCKs and HTML_BLOCKs. Tests
have not be written for either of them at this point, so I consider
them experimental at best.  Both **appear** to be working properly.

A parser for LINK_REF_DEFINITIONs has also been implemented. It is
VERY experimental, and no tests have been written for it either.
Currently it only supports link references that are defined entirely
on one line. This means the mandatory link label, mandatory
destination URL, and optional link title ALL MUST BE ON THE SAME LINE.
Otherwise, the block will most likely be parsed as a paragraph.

Two new static functions have been defined to pull some of the logic
out of the parse_block_type() function.  The first,
reset_parsing_variables() is the last operation completed (and it is
always completed) before returning to the function that called the
parsers [always build_queue()]. It updates and/or resets the global
variables that were defined to aid with the parsing based on the
outcome of parsing the current block. It also frees the reference
to the `ParseInfo` structure that has been created by the parsers.

The second new static function, check_last_block_type(), is always
the first operation called by the parsers. In other words, it is
called first, before any individual parsing function has been called.
It checks the mdblock_t of the last block parsed to determine if the
mdblock_t of the current block can be guessed. This is used to
monitor if we are currently *in* a block that needs a newline to
exit, such as FENCED_CODE_BLOCKs and HTML_BLOCKs. The function first
checks to make sure the line contains at least on character. It
returns NULL if it can make no such assumptions, and we get on our
merry parsing horse and ride into parsing-town.

~~~~~~~~~I'VE NEVER BEEN TO THIS PART OF PARSING-TOWN BEFORE~~~~~~~~~