fclose() files // rename test headers
Added fclose() calls to both the inputFile and outputFile in main().
Probably need to factor out a lot of the **teardown** code into a
separate function, but will work on that later.

Not closing the files were also causing an (OBVIOUS) memory leak that
I could not for the life of me figure out. Oh well. It is fixed now.

Also renamed the tests/test_parsing.h file to tests/tests_parsing.h.
This way it is more consistent with the other files in tests/.