Renamed utilities.c/h to be strings.c/h
This is mainly in keeping with what *utilities* are in this file. There are four functions in strings.c, and all four deal directly with creating and/or manipulating strings. They weren't kidding when they said strings are hard. All files (including Makefile) have been updated to reflect this renaming.