ADDED: Support for fenced code blocks
Two new mdblock_t's were added to handle the *fences* of these code blocks, FENCED_CODE_BLOCK_START and FENCED_CODE_BLOCK_END respectively. Both will carry a NULL value and are inserted into the queue only as placeholders to be used when printing.

The actual *body* of the code block will be parsed as a FENCED_CODE_BLOCK, and every line between the _START and _END mdblock_t will be appended the FENCED_CODE_BLOCK node with a newline separator. In other words, every fenced code block will be parsed exactly the same way, containing a maximum of three nodes:

    1. parse a FENCED_CODE_BLOCK_START
    2. parse all lines of input as a FENCED_CODE_BLOCK
    3. parse a FENCED_CODE_BLOCK_END

The tests still must be written for this type, I simply updated the array of string names of mdblock_t to reflect the two new enumerated values that were added.