Use new type string_t for strings instead of char*
The old readstring_t -- the type used to carry information from the file methods to the markdown methods -- has been renamed to string_t. This new type has been updated to hold three values:

1. char *string --> the raw string value
2. size_t size  --> the # of bytes allocated for *string
3. size_t len   --> the # of chars stored in *string

This new type allows for better dynamic memory management with the raw strings (lines) that are from the input file. The plan is to eventually update the markdown_t type to hold a string_t instead of a char *. Doing this would eliminate the repeated calls made to strlen(s), as we already know the size of the string, we read it from the file.