Fix create_substring() to count char's correctly
Previously the create_substring() function in strings.c was adding an extra character to specific mdblock_t's. Therefore, when a mdblock_t that would normally be returned as a NULL string (HORIZONTAL_RULE, BLANK_LINE, etc) was encountered, a string was allocated for exactly 1 character.

Part of the problem was located in the parsers, and I will now state the ideology of the create_substring() method once and for-all:

- start ==> the index of where the substring should start
- stop  ==> the index of where the substring should stop

This is an INCLUSIVE range.  In other words, given a string `s` of length *n*, the first character of the substring will be s[start], and the last character of the substring will be s[stop].  The NULL character will then be APPENDED to the end, as an extra character was added to the size in the function call to init_stringt().

NOTE: The NULL character should never be included in the range supplied to create_substring() -- aka, the character located at s[stop] should *never* be the NULL character. This is a bit harder, but necessary, given that not all substrings will include the end of the raw string read from file.

ANOTHER COROLLARY NOTE: When sending the `stop` value to create_substring(), great care should be taken to ensure that the value s->len -- the length member of the string_t node read from file -- *never* be used. THIS IS ALWAYS GUARANTEED TO BE THE NULL CHARACTER.  The value sent for `stop` should always be `s->len - 1`, or the index of the last **printable** character.

/endcharacterrant