Added rudimentary support for HTML blocks
This implementation requires that each HTML_BLOCK begin with a valid block-level HTML element. The list of elements were extracted from the CommonMark spec.  Once a line has been declared as an HTML block, all subsequent lines will become part of this HTML_BLOCK until a BLANK_LINE is encountered.  This BLANK_LINE effectively *clears* the parser, and new block-level elements are now parsed.

The major change from the CommonMark spec is the handling of when to end an HTML_BLOCK. The spec declares 8 different types of HTML blocks. Each of which begins and ends with different parameters. This seems to defeat the purpose of Markdown — HTML_BLOCK's should be used sparingly.  It wouldn't make sense unrolling and re-rolling all the logic it would take to determine which of the different 8 types of HTML_BLOCK we are entering, we are in, or we are exiting.

For this reason, and my general laziness, we will check the first line of an HTML_BLOCK for a valid block-level HTML element. If found, each subsequent line will be appended to this block until a BLANK_LINE is encountered.

See how much easier that is?  The logic could fit entirely in one sentence.