Removed Markdown elements from HTML block elements
If an element can be properly represented using some generic markdown syntax, and it is a block-level HTML element, it is excluded from being used to begin an HTML_BLOCK.

So, you cannot using a <p> tag to begin an HTML_BLOCK. Because you can already represent a paragraph in Markdown with considerably less effort than typing the tag.. you must use the Markdown syntax to represent paragraphs.

Elements that this pertains to:
- <p>
- <h1>, <h2>, <h3>, <h4>, <h5>, <h6>
- <blockquote>
- <ul>, <ol>, <li>
- <hr>

This will greatly shorten the time it takes to compare valid HTML_BLOCK tags with the ones parsed from the input file.