Add specific mdblock_t data to markdown_t nodes
This is the first in a slew of new updates to make the parser operate in a more *recursive-descent* manner. The idea is to add a void (generic) pointer "data" member to every markdown_t node. Most nodes will ignore this member, just initialize it to NULL and then forget about it.  But specific, dare-i-say more *complicated*, nodes will be able to store more state information than just a string in the markdown_t node.

Previously, I was creating hack-y ways around storing additional state information. Now, if we are initializing the markdown_t node to be one of the *special* nodes — currently: fenced code blocks, blockquotes, lists, and link references — it will initialize a separate structure specific to that mdblock_t.

For example, we encounter a fenced code block by stumbling up its opening fence and *info string*. We will do what we normally do, parse the fence to ensure that it meets all the specifications.  Then, instead of storing the info string in the "value" portion of the markdown_t node, we will initialize another node.  In this case it would be a md_code_block_t node. This node has two members:

1. char fence; // fence character used
2. string_t *info; // info string from opening fence

The md_code_block_t node will then be assigned to the void (generic) pointer on the markdown_t node.

Normally, using void pointers in this way would cause a huge fucking headache because anytime you would want to access that information, you have to cast the void pointer to the correct type, or forever live in fear of garbage values. But in this case, we already know the type — it can be easily determined from the mdblock_t!

I am sure there will be unforeseen consequences to this change. But I'd rather continue to revel in the fucking glory that is void pointers.