Begin parser rewrite to be recursive-descent
This is the first in a long string of rewrites to the parser. For now, only PARAGRAPH, SETEXT_HEADING_x, and BLANK_LINE are working.

The overall goal is to make the parser be more self-sufficient. Previously, when the parser would return the markdown_t node back to the markdown-queue methods, there was still parsing-related-logic being sorted out.

Using a more recursive-descent approach allows the parser to make the logical decisions before it sends the node back to the calling function. Here is the main example (that has been implemented in this commit):

Once a PARAGRAPH has been parsed, the parser will continue to consume and parse raw lines of input until it is sure that we have reached the end of this PARAGRAPH node. It will also check the first non-PARAGRAPH raw line of input to determine if this is a line that will convert the previously parsed PARAGRAPH into a SETEXT_HEADING_x.

This can now all be done with a single call to the parser. We no longer call the parser with a line of input; instead we call it with a file pointer, and it will consume its own line(s) of input.