Change read_line() to detect EOF and return NULL
The read_line() function will now detect when EOF is reached, and it will return NULL if no characters were read from file.

The markdown() function -- which operates effectively as the parsing loop (for now) -- was updated to reflect this change in logic.

This is the first step on the way to make the parser independent of the markdown() function. Eventually, the goal is to move the markdown() function
into the parsers.c file, and have it be the external API for the entire parser -- then the markdown.c file will only contain the methods for operating on the data structures created by the parser.