REMOVE: block_types.h file
The md_code_block_t structure has been moved to markdown.h to better
reflect the nature of the structure. It has also been renamed to
`CodeBlock`.

All of the type declarations in all of the files have been updated
to compile after this change. All of the import statements have
also been removed, importing markdown.h is now enough.