ADD: new error function throw_invalid_output_type()
This function is called when the user provides an output type that
patdown cannot recognize, even after it attempts to *normalize*
their input.

This also creates a new class of errors wherein the error function
is thrown -- or *called* -- and the program doesn't exit, but rather,
a dialog is printed describing the error and control is returned to
the caller.