UPDATE: Use uint8_t instead of char* for String's
This is a major change in the way strings are handled inside the
program. It was redesigned to be UTF-8 friendly. Instead of storing
strings as a char* array, we now stored, read, and print them as
uint8_t arrays.

All of the String API has been updated to reflect this change.