Change queue implementation to be file-private
This is a major difference from the previous iteration. In so
many words, the entire implementation of the Markdown queue is now
static to the markdown.c file. This includes the `Markdown` nodes,
as they are no longer needed by any other files.

An set of functions will be implemented in order to add/free/get
the data from the queue.