Overhaul the parser to use new String API
This is just the basics of parsing under the new String API. No more
reading the file line-by-line and appending lines to a giant buffer,
now we work with just 1 giant buffer, the raw bytes read from the
file.

So far the markdown(), block_parser(), and is_blank_line() functions
are working. The parse_paragraph() function is half-working, it just
needs to check the next line for a lazy continuation.

This is good -- we are (mostly) UTF-8 compatible!