Add ATX headers are now UTF-8 compliant
Any line probable of being an ATX header should call is_atx_header(),
which will confirm or deny this. If confirmed, it will call
parse_atx_header() which will add a node to the queue. If this line
is found to not be an ATX header, -1 will be returned from
is_atx_header().