Initialize Markdown nodes so that String is never NULL
Instead, just take advantage of the "convenient initializer"
(I'm borrowing Swift terminology) for String's by calling
init_string(0). This will create an "empty String" -- wherein
length and allocd are set to 0, and data is set to NULL.

This just makes traversing the queue a hell of a whole lot easier.
Now you can always be sure that the String node will exist, no need
to do a fuck-ton of NULL checking.