Add the ability to just check syntax on parsers
This adds a Boolean parameter to the following parsers:

- is_blank_line()
- is_paragraph()
- is_atx_header()
- is_horizontal_rule()

Basically all the parsers of the form `is_block()`. This allows other
parsing functions that must check for continuation among multiple
lines to call these functions safely (i.e. without adding a node(s)
to the queue). When called with their boolean argument set to false,
they will just check the syntax of the block in question -- returning
-1 if this is a different block, or a positive integer if correct.

Two constants have been defined to add semantics to these new calls:

- PARSE_BLK is defined as true
- CHK_SYNTX is defined as false

These should be used when making the function calls to state the
purpose of the call explicitly.