Make get_last_block() return the current block's type
This is a bit of a hack, and a better solution may find it's way
into the code eventually, but it will do for now. Basically, when
parsing a (potentially) multiline block like a paragraph, we need
PARAGRAPH to be returned from get_last_block() calls in order for
the other parsing functions to believe that they are following a
paragraph -- even though the PARAGRAPH block hasn't been added to
the queue yet.

This is solved via the set_current_block() function. It sets a static
variable to the current block, and that value is returned from
subsequent calls to get_last_block() -- until a call to add_markdown()
if made, and the current block value is reset to UNKNOWN.