Fix block parser to finish parsing at the right time
The fix involves keeping track of the number of bytes the parser
has consumed, and comparing that to the number of bytes in the
input string in order to determine when we are done parsing.

This solves the majority of the problems with making logical
decisions based on uninitialized values, i.e. calling a parsing
function (or count_indentation()) on (or past) the end of the
input string.

There is still an edge case wherein if the last block in the input
string is a paragraph, the paragraph parser will call the blank
line parser on the end of the input string. As of now I haven't
found an elegant solution to this problem...