Fix error in testing script
All of the tests are conducted using bash's string equality
operator `==` -- which is the same this as `=` in bash (jesus christ).
Previously, because the right-hand operator was not string-escaped,
it was using pattern matching instead of literal matching.

The key here is to always string escape everything in bash when
doing comparisons. Think it like copy-on-write in objective-c.