Add HTML Block support for types 1 - 5
These blocks all are closed when an end tag is encountered, which
is accomplished by a new function line_contains_endtag() -- which
takes our bag of UTF8 bytes and a substring and search for the next
occurrence of that substring. It returns true if that occurrence
is on the current line, false otherwise.

When parsing these blocks, we just copy every line until
line_contains_endtag() returns true, then copy one last line (the
line that contains the end tag), and add the node to the queue.