Improve the link reference BST implementation
The LinkRef-type is still made public, but the actual implementation
of the BST is private to links.c. This better reflects the Markdown
queue created in markdown.c.

Because the type itself is still public, the parser is expected to
actually fill the nodes -- add the labels, destinations, and titles
to each node. It then calls insert_link_ref() to add this link to
the private BST. Finally, it calls add_markdown() with a
LINK_REF_DEF and the LinkRef node as the addt_info parameter. This
way the source file is still correctly represented in the
sequential queue we build.