Fix blockquote parsing to handle lazy continuation
This commit actuals fixes the issue -- basically a rewrite of the
previous two commits: eb78d83 and 3d97ba1. I guess the third times
the charm.

Every time a blockquote is closed -- a line is reached that does not
begin with the required '>' character -- the content that we have
stripped of the leading characters is parsed. The next line (if it
exists) is examined to determine if it would be a lazy paragraph
continuation. If thats true and the last block inside the blockquote
was a paragraph, we pop that block and continue parsing before
closing the blockquote. Otherwise, closing the blockquote and return
to the block parser.

Woooohhhhhhhh. I need a beer.